Add type cast to resolve C4804 (unsafe bool usage) warning in MSVC. Closes #2731.

+++ Cython/Utility/ModuleSetupCode.c
@@ -207,7 +207,7 @@
   #undef MASK
   /* Compile-time sanity check that these are indeed equal.  Github issue #2670. */
   #ifdef SIZEOF_VOID_P
-    enum { __pyx_check_sizeof_voidp = 1/(SIZEOF_VOID_P == sizeof(void*)) };
+    enum { __pyx_check_sizeof_voidp = 1 / (int)(SIZEOF_VOID_P == sizeof(void*)) };
   #endif
 #endif
